BE/auth/requirements.txt
Django>=4.2.0
psycopg[binary]
djangorestframework
django-filter
drf-yasg
django-cors-headers
requests
